studio/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pydantic>=2.7.0

# Audio I/O
soundfile>=0.12.1
pydub>=0.25.1
librosa>=0.10.1

# Stem separation
demucs>=4.0.1

# Basic Pitch
basic-pitch>=0.3.3

# MIDI
pretty_midi>=0.2.10
midiutil>=1.2.1

# FluidSynth
pyfluidsynth>=1.3.3

# Effects chain
pedalboard>=0.9.0

# AI mastering
matchering>=2.0.6

# HuggingFace / MusicGen
transformers>=4.40.0
torch>=2.2.0
torchaudio>=2.2.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
numpy>=1.26.0
